pc/pyproject.toml
[project]
name = "pc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "protobuf>=6.32.1",
    "pyserial>=3.5",
]
